server/requirements.txt
Django
djangorestframework
minio
psycopg[binary]